Enabled CSRF-protection for the web application and the dropzone part, which required some changes regarding the CSP as well.
It was not possible to have the dropzone JS code in the static helper.js file and have a CSRF token enabled as well. Therefore the dropzone.config code was moved back into the template and the CSP nonce value was passed to dropzone.configure(), which required a change in the flask-dropzone code (https://github.com/greyli/flask-dropzone/pull/40/).